+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Patterns now show the data needed to render them
 - Removed `Pattern::composer`
 - Added `View::composer($name, $callable)` as a syntactic sugar for `Event::listen("view.$name", $callable)`
+- Primer specific data moved to `primer` namespace in `data.json` files
 
 ### Breaking
 - New Handlebars engine drops support for `.handlebars` extensions, only `.hbs` now supported out of the box